price for custom

+++ web_bakery_marketplace_BE/src/services/checkout.service.ts
@@ -99,11 +99,12 @@ class CheckoutService {
                 for (const orderProductID of order.order_products) {
                     const orderProduct = await orderProductRepo.getOderProductById(orderProductID.toString());
                     if (orderProduct) {
-                        const updateOderProduct = await orderProductRepo.updateOderProduct(orderProduct._id.toString(), { status: 'confirmed' });
-                        console.log("updateOderProduct", updateOderProduct)
+                        const updateOderProduct = await orderProductRepo.updateOderProduct(orderProduct._id.toString(), { status: 'success' });
                         order_products.push(updateOderProduct);
                         //payment success, remove product in cart
-                        await cartRepo.removeProductFromCart(order.user_id.toString(), orderProduct.product_id.toString());
+                        if (!orderProduct.isCustomCake && orderProduct.product_id) {
+                            await cartRepo.removeProductFromCart(order.user_id.toString(), orderProduct.product_id.toString());
+                        }
                     }
                 }
 
@@ -120,8 +121,10 @@ class CheckoutService {
                     const orderProduct = await orderProductRepo.getOderProductById(orderProductID.toString());
                     if (orderProduct) {
                         const quantity = orderProduct.quantity;
-                        const updateInventory = await inventoryRepo.updateInventory(orderProduct.product_id.toString(), quantity);
-                        console.log("updateInventory", updateInventory)
+                        if (!orderProduct.isCustomCake && orderProduct.product_id) {
+                            const updateInventory = await inventoryRepo.updateInventory(orderProduct.product_id.toString(), quantity);
+                            console.log("updateInventory", updateInventory)
+                        }
                         await orderProductRepo.deleteOderProduct(orderProduct._id.toString());
                     }
                 }
